@@ -19,21 +19,13 @@ namespace ecoPlanerWeb
             Console.WriteLine("Hello World!");
             ResourceTypes.Init();
 
-            var op = new Region(true);
-            var op2 = new Region(false);
-            var regions = new List<Region> {
-                op,
-                op2
-            };
-            op.ConnectNeighbour(op2);
+
 
             int iter = 0;
             while (true)
             {
                 Console.WriteLine("------------------------New gen------------------------------------" + iter++);
-                regions.ForEach(re => re.Update());
 
-                regions.ForEach(re => re.CleanUp());
                 System.Threading.Thread.Sleep(10);
             }
         }
